// src/app/main/budget/shared/Dtos/restriction-dto.ts
export class RestrictionDto {
  isActive: boolean = true;
  name: string;
  voucherPrefix: string;
  isCreationAllowed: boolean = true;
  creationRestrictionDate: Date;
  isEditAllowed: boolean = true;
  editRestrictionDate: Date;
  isApprovalAllowed: boolean = true;
  approvalRestrictionDate: Date;
  isUnapprovalAllowed: boolean = true;
  unapprovalRestrictionDate: Date;
  isRevisionAllowed: boolean = true;
  revisionRestrictionDate: Date;
}